requirements.txt
pytest
asyncio
openai>=1.3.0
python-dotenv
pydantic>=2.0.0
